Changed Condition for Years to False
Changed each condition to false that triggered the year drop down (since these were originally used for subcat_id =1 and =2, but we changed to =2 and =3 as a temporary solution before we realized we needed to remove those items altogether. Left comments explaining what the previous condition was and why we removed them.